+++ __tests__/services/connection.service.spec.ts
@@ -5,9 +5,11 @@ import { v4 } from 'uuid';
 
 import {
   ActorType,
+  ControlEventAction,
   DataChunkDataType,
   InworldPacket as ProtoPacket,
 } from '../../proto/packets.pb';
+import { TtsPlaybackAction } from '../../src/common/data_structures';
 import { protoTimestamp } from '../../src/common/helpers';
 import {
   CHAT_HISTORY_TYPE,
@@ -29,6 +31,7 @@ import { WorldEngineService } from '../../src/services/pb/world_engine.service';
 import {
   capabilitiesProps,
   convertAgentsToCharacters,
+  convertPacketFromProto,
   createAgent,
   generateSessionToken,
   previousState,
@@ -41,6 +44,7 @@ import {
 const onError = jest.fn();
 const onMessage = jest.fn();
 const onDisconnect = jest.fn();
+const onInterruption = jest.fn();
 const agents = [createAgent(), createAgent()];
 const characters = convertAgentsToCharacters(agents);
 const scene = { agents };
@@ -408,15 +412,60 @@ describe('open manually', () => {
 
     expect(open).toHaveBeenCalledTimes(1);
   });
+
+  test('should add previous state packets to history', async () => {
+    jest
+      .spyOn(WorldEngineService.prototype, 'loadScene')
+      .mockImplementationOnce(() =>
+        Promise.resolve({
+          ...scene,
+          previousState: {
+            stateHolders: [
+              {
+                packets: [incomingTextEvent],
+              },
+            ],
+          },
+        }),
+      );
+
+    connection = new ConnectionService({
+      name: SCENE,
+      config: {
+        connection: { autoReconnect: false, gateway: { hostname: '' } },
+        capabilities: capabilitiesProps,
+        history: { previousState: true },
+      },
+      user,
+      onError,
+      onMessage,
+      onDisconnect,
+      onHistoryChange: (history: HistoryItem[]) => {
+        expect(history).toEqual([convertPacketFromProto(incomingTextEvent)]);
+      },
+      grpcAudioPlayer,
+      generateSessionToken,
+      webRtcLoopbackBiDiSession,
+    });
+
+    await connection.openManually();
+  });
 });
 
 describe('send', () => {
+  let server: WS;
+  const HOSTNAME = 'localhost:1234';
+
   let connection: ConnectionService;
 
   const onHistoryChange = jest.fn();
 
   beforeEach(() => {
     jest.clearAllMocks();
+
+    server = new WS(`ws://${HOSTNAME}/v1/session/default`, {
+      jsonProtocol: true,
+    });
 
     connection = new ConnectionService({
       name: SCENE,
@@ -428,11 +477,17 @@ describe('send', () => {
       onError,
       onMessage,
       onDisconnect,
+      onInterruption,
       onHistoryChange,
       grpcAudioPlayer,
       generateSessionToken,
       webRtcLoopbackBiDiSession,
     });
+  });
+
+  afterEach(() => {
+    server.close();
+    WS.clean();
   });
 
   test('should throw error in case of connection is inactive on send call', async () => {
@@ -519,6 +574,8 @@ describe('send', () => {
   });
 
   test('should interrupt on text sending', async () => {
+    const interactionId = v4();
+    const utteranceId = v4();
     const write = jest
       .spyOn(WebSocketConnection.prototype, 'write')
       .mockImplementation(writeMock);
@@ -536,8 +593,8 @@ describe('send', () => {
           ...audioEvent,
           packetId: {
             packetId: audioEvent.packetId.packetId,
-            interactionId: v4(),
-            utteranceId: v4(),
+            interactionId,
+            utteranceId,
           },
         }),
       ]);
@@ -547,6 +604,76 @@ describe('send', () => {
     expect(open).toHaveBeenCalledTimes(1);
     expect(write).toHaveBeenCalledTimes(2);
     expect(cancelResponse).toHaveBeenCalledTimes(1);
+    expect(onInterruption).toHaveBeenCalledTimes(1);
+    expect(onInterruption).toHaveBeenCalledWith({
+      interactionId,
+      utteranceId: [utteranceId],
+    });
+  });
+
+  test('should add playback mute event to queue in case of auto reconnect', async () => {
+    const open = jest.spyOn(ConnectionService.prototype, 'open');
+    const write = jest
+      .spyOn(WebSocketConnection.prototype, 'write')
+      .mockImplementationOnce(writeMock);
+    const wsOpen = jest
+      .spyOn(WebSocketConnection.prototype, 'open')
+      .mockImplementationOnce(jest.fn());
+    jest
+      .spyOn(WorldEngineService.prototype, 'loadScene')
+      .mockImplementationOnce(() => Promise.resolve(scene));
+    jest
+      .spyOn(ConnectionService.prototype, 'getTtsPlaybackAction')
+      .mockImplementationOnce(() => TtsPlaybackAction.MUTE);
+
+    connection = new ConnectionService({
+      name: SCENE,
+      config: {
+        connection: { gateway: { hostname: '' } },
+        capabilities: capabilitiesProps,
+      },
+      user,
+      grpcAudioPlayer,
+      generateSessionToken,
+      webRtcLoopbackBiDiSession,
+    });
+
+    await connection.send(() => textEvent);
+
+    expect(open).toHaveBeenCalledTimes(1);
+    expect(write).toHaveBeenCalledTimes(1);
+    expect(wsOpen.mock.calls[0][0].packets?.length).toEqual(1);
+    expect(
+      wsOpen.mock.calls[0][0].packets[0].getPacket().control?.action,
+    ).toEqual(ControlEventAction.TTS_PLAYBACK_MUTE);
+  });
+
+  test('should not add playback mute event to queue in case of manual reconnect', async () => {
+    connection = new ConnectionService({
+      name: SCENE,
+      config: {
+        connection: { gateway: { hostname: HOSTNAME }, autoReconnect: false },
+        capabilities: capabilitiesProps,
+      },
+      user,
+      grpcAudioPlayer,
+      generateSessionToken,
+      webRtcLoopbackBiDiSession,
+    });
+
+    const wsOpen = jest.spyOn(WebSocketConnection.prototype, 'open');
+
+    jest
+      .spyOn(WorldEngineService.prototype, 'loadScene')
+      .mockImplementationOnce(() => Promise.resolve(scene));
+    jest
+      .spyOn(ConnectionService.prototype, 'getTtsPlaybackAction')
+      .mockImplementationOnce(() => TtsPlaybackAction.MUTE);
+
+    await connection.open();
+    await server.connected;
+
+    expect(wsOpen.mock.calls[0][0].packets).toEqual(undefined);
   });
 });
 
@@ -573,6 +700,7 @@ describe('onMessage', () => {
       onError,
       onMessage,
       onDisconnect,
+      onInterruption,
       grpcAudioPlayer,
       generateSessionToken,
       webRtcLoopbackBiDiSession,