+++ src/common/data_structures.ts
@@ -56,16 +56,22 @@ export interface ConnectionConfig {
   disconnectTimeout?: number;
   gateway?: Gateway;
 }
+
+export interface HistoryConfig {
+  previousState?: boolean;
+}
 export interface ClientConfiguration {
   connection?: ConnectionConfig;
   capabilities?: Capabilities;
   audioPlayback?: AudioPlaybackConfig;
+  history?: HistoryConfig;
 }
 
 export interface InternalClientConfiguration {
   connection?: ConnectionConfig;
   capabilities: CapabilitiesRequest;
   audioPlayback?: AudioPlaybackConfig;
+  history?: HistoryConfig;
 }
 
 export interface CancelResponses {
@@ -95,6 +101,12 @@ export enum AudioSessionState {
   UNKNOWN = 'UNKNOWN',
   START = 'START',
   END = 'END',
+}
+
+export enum TtsPlaybackAction {
+  UNKNOWN = 'UNKNOWN',
+  MUTE = 'MUTE',
+  UNMUTE = 'UNMUTE',
 }
 
 export interface Extension<InworldPacketT, HistoryItemT> {