+++ src/services/connection.service.ts
@@ -1,5 +1,8 @@
 import { InworldPacket as ProtoPacket } from '../../proto/packets.pb';
-import { LoadSceneResponseAgent } from '../../proto/world-engine.pb';
+import {
+  LoadSceneResponseAgent,
+  PreviousState,
+} from '../../proto/world-engine.pb';
 import { ClientRequest, LoadSceneResponse } from '../../proto/world-engine.pb';
 import {
   AudioSessionState,
@@ -10,6 +13,7 @@ import {
   Extension,
   GenerateSessionTokenFn,
   InternalClientConfiguration,
+  TtsPlaybackAction,
   User,
 } from '../common/data_structures';
 import {
@@ -22,6 +26,7 @@ import { GrpcWebRtcLoopbackBiDiSession } from '../components/sound/grpc_web_rtc_
 import { Player } from '../components/sound/player';
 import {
   Connection,
+  QueueItem,
   WebSocketConnection,
 } from '../connection/web-socket.connection';
 import { Character } from '../entities/character.entity';
@@ -42,6 +47,7 @@ interface ConnectionProps<InworldPacketT, HistoryItemT> {
   onError?: (err: Event | Error) => Awaitable<void>;
   onMessage?: (packet: InworldPacketT) => Awaitable<void>;
   onDisconnect?: () => Awaitable<void>;
+  onInterruption?: (props: CancelResponsesProps) => Awaitable<void>;
   onHistoryChange?: (history: HistoryItem[]) => Awaitable<void>;
   grpcAudioPlayer: GrpcAudioPlayback;
   webRtcLoopbackBiDiSession: GrpcWebRtcLoopbackBiDiSession;
@@ -56,6 +62,7 @@ export class ConnectionService<
   private player = Player.getInstance();
   private state: ConnectionState = ConnectionState.INACTIVE;
   private audioSessionAction = AudioSessionState.UNKNOWN;
+  private ttsPlaybackAction = TtsPlaybackAction.UNKNOWN;
 
   private scene: LoadSceneResponse;
   private session: SessionToken;
@@ -170,9 +177,12 @@ export class ConnectionService<
       if (this.state === ConnectionState.LOADED) {
         this.state = ConnectionState.ACTIVATING;
 
+        const packets = this.getPacketsToSentOnOpen();
+
         await this.connection.open({
           session: this.session,
           convertPacketFromProto: this.extension.convertPacketFromProto,
+          ...(packets.length && { packets }),
         });
 
         this.scheduleDisconnect();
@@ -202,6 +212,14 @@ export class ConnectionService<
 
   getAudioSessionAction() {
     return this.audioSessionAction;
+  }
+
+  setTtsPlaybackAction(action: TtsPlaybackAction) {
+    this.ttsPlaybackAction = action;
+  }
+
+  getTtsPlaybackAction() {
+    return this.ttsPlaybackAction;
   }
 
   async interrupt() {
@@ -307,6 +325,10 @@ export class ConnectionService<
           client,
         });
 
+        if (this.connectionProps?.config.history?.previousState) {
+          this.setPreviousState(this.scene?.previousState);
+        }
+
         await this.loadCharactersList();
       }
 
@@ -350,6 +372,24 @@ export class ConnectionService<
         this.connectionProps.config.connection.disconnectTimeout,
       );
     }
+  }
+
+  private setPreviousState(previousState: PreviousState) {
+    const { stateHolders = [] } = previousState || {};
+
+    stateHolders.forEach((stateHolder) => {
+      stateHolder.packets?.forEach((packet) =>
+        this.history.addOrUpdate({
+          grpcAudioPlayer: this.connectionProps.grpcAudioPlayer,
+          characters: this.characters,
+          packet: this.extension.convertPacketFromProto(packet),
+        }),
+      );
+
+      if (stateHolders.length) {
+        this.connectionProps.onHistoryChange?.(this.getHistory());
+      }
+    });
   }
 
   private cancelScheduler() {
@@ -507,10 +547,14 @@ export class ConnectionService<
         [cancelReponses.interactionId]: true,
       };
 
-      this.history.filter({
+      const interruptionData = {
         utteranceId: cancelReponses.utteranceId ?? [],
         interactionId: cancelReponses.interactionId,
-      });
+      };
+
+      this.connectionProps.onInterruption?.(interruptionData);
+
+      this.history.filter(interruptionData);
     }
   }
 
@@ -556,5 +600,19 @@ export class ConnectionService<
     });
 
     this.intervals = [];
+  }
+
+  private getPacketsToSentOnOpen() {
+    const packets: QueueItem<InworldPacketT>[] = [];
+
+    if (this.isAutoReconnected()) {
+      if (this.getTtsPlaybackAction() === TtsPlaybackAction.MUTE) {
+        packets.push({
+          getPacket: () => this.getEventFactory().ttsPlaybackMute(true),
+        });
+      }
+    }
+
+    return packets;
   }
 }